bundler: spread relocated types across their owning modules; add splitting regression test
Distribute ungate_support's former contents to the files that own the
data instead of concentrating them in bundle_v2.rs + lib.rs:

- CrossChunkImport/CrossChunkImportItem + the cross_chunk_import alias and
  the bun_renamer module (ChunkRenamer) -> Chunk.rs, where Chunk.renamer
  and Chunk.imports_from_other_chunks live
- EntryPoint and js_meta (JSMeta/ImportData/ExportData/...) -> LinkerGraph.rs,
  the graph metadata they describe (LinkerGraph.entry_points / .meta)
- the html_import_manifest calling-convention adapter -> HTMLImportManifest.rs,
  next to the module it wraps
- the bundle_v2.zig top-level value types (PartRange, StableRef, ImportTracker,
  CompileResult, ContentHasher, WrapKind, DeclInfo, generic_path_with_pretty_initialized,
  target_from_hashbang) stay in bundle_v2.rs
- lib.rs keeps only the crate-name aliases/glue and re-exports the above from
  their new homes, so every crate::* spelling callers use is unchanged

Add regression/SplittingWithSharedChunkAndCSS, which bundles with code
splitting + a shared chunk + CSS so the cross-chunk import, chunk hashing,
cross-chunk renaming, and CSS-chunk paths all run end-to-end.
