+++ src/bundler/Chunk.rs
@@ -20,7 +20,6 @@ use bun_sourcemap as source_map;
 use crate::analyze_transpiled_module;
 use crate::bun_css;
 use crate::bun_fs;
-use crate::bun_renamer;
 
 use crate::Graph::Graph;
 use crate::html_import_manifest as HTMLImportManifest;
@@ -1520,9 +1519,68 @@ impl<'a, 'ctx> fmt::Display for CssImportOrderDebug<'a, 'ctx> {
     }
 }
 
-pub(crate) type ImportsFromOtherChunks =
-    ArrayHashMap<IndexInt, crate::cross_chunk_import::ItemList>;
-// TODO(port): CrossChunkImport.Item.List — assuming exported as ItemList from cross_chunk_import module
+pub(crate) type ImportsFromOtherChunks = ArrayHashMap<IndexInt, cross_chunk_import::ItemList>;
+
+#[derive(Default, Clone)]
+pub struct CrossChunkImportItem {
+    pub export_alias: Box<[u8]>,
+    pub r#ref: Ref,
+}
+pub type CrossChunkImportItemList = Vec<CrossChunkImportItem>;
+#[derive(Default)]
+pub struct CrossChunkImport {
+    pub chunk_index: IndexInt,
+    pub sorted_import_items: core::mem::ManuallyDrop<CrossChunkImportItemList>,
+}
+
+pub mod cross_chunk_import {
+    pub(crate) type ItemList = super::CrossChunkImportItemList;
+}
+
+impl CrossChunkImportItem {
+    pub fn less_than(_: (), a: &CrossChunkImportItem, b: &CrossChunkImportItem) -> bool {
+        strings::order(&a.export_alias, &b.export_alias) == core::cmp::Ordering::Less
+    }
+}
+
+impl CrossChunkImport {
+    pub fn less_than(_: (), a: &CrossChunkImport, b: &CrossChunkImport) -> bool {
+        a.chunk_index < b.chunk_index
+    }
+
+    pub fn sorted_cross_chunk_imports(
+        list: &mut Vec<CrossChunkImport>,
+        chunks: &mut [Chunk],
+        imports_from_other_chunks: &mut ImportsFromOtherChunks,
+    ) -> Result<(), bun_core::Error> {
+        list.clear();
+        list.reserve(imports_from_other_chunks.count());
+
+        for i in 0..imports_from_other_chunks.count() {
+            let chunk_index = imports_from_other_chunks.keys()[i];
+            let chunk = &mut chunks[chunk_index as usize];
+
+            let exports_to_other_chunks = &chunk.content.javascript().exports_to_other_chunks;
+            let import_items = &mut imports_from_other_chunks.values_mut()[i];
+            for item in import_items.slice_mut() {
+                item.export_alias = (*exports_to_other_chunks.get(&item.r#ref).unwrap()).into();
+                debug_assert!(!item.export_alias.is_empty());
+            }
+            import_items
+                .slice_mut()
+                .sort_by(|a, b| strings::order(&a.export_alias, &b.export_alias));
+
+            list.push(CrossChunkImport {
+                chunk_index,
+                sorted_import_items: import_items.shallow_copy(),
+            });
+        }
+
+        list.sort_by_key(|a| a.chunk_index);
+        Ok(())
+    }
+}
+
 // `Chunk` is bump-arena-allocated (no Drop on free); boxing the large arm
 // would leak. The CSS/JS chunk size diff is acceptable.
 #[allow(clippy::large_enum_variant)]
@@ -1577,5 +1635,34 @@ impl Content {
 pub use crate::DeferredBatchTask::DeferredBatchTask;
 pub use crate::ParseTask;
 pub use crate::ThreadPool;
+
+pub mod bun_renamer {
+    pub use bun_js_printer::renamer::*;
+
+    #[derive(Default)]
+    pub enum ChunkRenamer {
+        #[default]
+        None,
+        Number(Box<bun_js_printer::renamer::NumberRenamer>),
+        Minify(Box<bun_js_printer::renamer::MinifyRenamer>),
+    }
+
+    impl ChunkRenamer {
+        pub(crate) fn name_for_symbol(&mut self, ref_: bun_ast::Ref) -> &[u8] {
+            match self {
+                ChunkRenamer::None => unreachable!("ChunkRenamer not initialized"),
+                ChunkRenamer::Number(r) => r.name_for_symbol(ref_),
+                ChunkRenamer::Minify(r) => r.name_for_symbol(ref_),
+            }
+        }
+        pub(crate) fn as_renamer(&mut self) -> bun_js_printer::renamer::Renamer<'_, '_> {
+            match self {
+                ChunkRenamer::None => unreachable!("ChunkRenamer not initialized"),
+                ChunkRenamer::Number(r) => bun_js_printer::renamer::Renamer::NumberRenamer(r),
+                ChunkRenamer::Minify(r) => bun_js_printer::renamer::Renamer::MinifyRenamer(r),
+            }
+        }
+    }
+}
 
 // ported from: src/bundler/Chunk.zig
