+++ src/bundler/HTMLImportManifest.rs
@@ -349,4 +349,43 @@ pub fn write<W: Write + ?Sized>(
     Ok(())
 }
 
+pub mod html_import_manifest {
+    use crate::Graph::Graph;
+    use crate::{LinkerGraph, chunk::Chunk};
+
+    pub use super::{EscapedJson, HTMLImportManifest};
+
+    #[inline]
+    pub(crate) fn format_escaped_json<'a>(
+        index: u32,
+        graph: &'a Graph,
+        chunks: &'a [Chunk],
+        linker_graph: &'a LinkerGraph,
+    ) -> EscapedJson<'a> {
+        super::HTMLImportManifest {
+            index,
+            graph,
+            chunks,
+            linker_graph,
+        }
+        .format_escaped_json()
+    }
+
+    pub fn write_escaped_json(
+        index: u32,
+        graph: &Graph,
+        linker_graph: &LinkerGraph<'_>,
+        chunks: &[Chunk],
+        w: &mut &mut [u8],
+    ) -> Result<(), core::fmt::Error> {
+        let taken = core::mem::take(w);
+        let mut fbs = bun_io::FixedBufferStream::new_mut(taken);
+        super::write_escaped_json(index, graph, linker_graph, chunks, &mut fbs)
+            .map_err(|_| core::fmt::Error)?;
+        let bun_io::FixedBufferStream { buffer, pos } = fbs;
+        *w = &mut buffer[pos..];
+        Ok(())
+    }
+}
+
 // ported from: src/bundler/HTMLImportManifest.zig
