+++ src/bundler/LinkerGraph.rs
@@ -10,10 +10,7 @@ use bun_collections::{AutoBitSet, DynamicBitSetUnmanaged as BitSet, MultiArrayLi
 use bun_core::PathString;
 
 use crate::IndexStringMap::IndexStringMap;
-use crate::{
-    EntryPoint, ImportTracker, Index, JSAst, JSMeta, Part, Ref, ResolvedExports,
-    TopLevelSymbolToParts, UseDirective, entry_point, import_record, index, js_meta, part,
-};
+use crate::{ImportTracker, Index, JSAst, Part, Ref, UseDirective, import_record, index, part};
 // `items_<field>()` column accessors — bring the `*ListExt` traits into scope.
 // PORT NOTE: `BundledAstColumns` is emitted by ``
 // on `BundledAst`; un-gating here is paired with that derive landing in
@@ -21,11 +18,173 @@ use crate::{
 // already imports as `BundledAstField`).
 bun_core::declare_scope!(LinkerGraph, visible);
 
-// PORT NOTE: MultiArrayList per-field column accessors are now generated by
-// `` as `*ListExt` traits with
-// `items_<field>()` / `items_<field>_mut()`. The gated method bodies below
-// were drafted against an older `.items().field_name` shape; rewritten to the
-// `items_<field>()` spelling (matches `LinkerContext.rs`).
+pub mod entry_point {
+    use bun_collections::MultiArrayList;
+    use bun_core::PathString;
+
+    #[derive(Default)]
+    pub struct EntryPoint {
+        pub output_path: PathString,
+        pub source_index: crate::IndexInt,
+        pub output_path_was_auto_generated: bool,
+    }
+
+    pub type List = MultiArrayList<EntryPoint>;
+
+    bun_collections::multi_array_columns! {
+        pub trait EntryPointColumns for EntryPoint {
+            output_path: PathString,
+            source_index: crate::IndexInt,
+            output_path_was_auto_generated: bool,
+        }
+    }
+
+    impl EntryPoint {
+        pub type Kind = Kind;
+    }
+
+    #[repr(u8)]
+    #[derive(Clone, Copy, PartialEq, Eq, Default)]
+    pub enum Kind {
+        #[default]
+        None,
+        UserSpecified,
+        DynamicImport,
+        Html,
+    }
+    impl Kind {
+        #[inline]
+        pub fn is_entry_point(self) -> bool {
+            self != Self::None
+        }
+        #[inline]
+        pub fn is_user_specified_entry_point(self) -> bool {
+            self == Self::UserSpecified
+        }
+        #[inline]
+        pub fn is_server_entry_point(self) -> bool {
+            self == Self::UserSpecified
+        }
+        #[inline]
+        pub fn output_kind(self) -> crate::options::OutputKind {
+            match self {
+                Self::UserSpecified => crate::options::OutputKind::EntryPoint,
+                _ => crate::options::OutputKind::Chunk,
+            }
+        }
+    }
+}
+
+pub mod js_meta {
+    use bun_alloc::{AstAlloc, AstVec};
+    use bun_ast::{Dependency, Ref};
+    use bun_collections::array_hash_map::StringContext;
+    use bun_collections::{ArrayHashMap, AutoContext, StringArrayHashMap};
+
+    use crate::{ImportTracker, Index, WrapKind};
+
+    pub struct ImportData {
+        pub re_exports: AstVec<Dependency>,
+        pub data: ImportTracker,
+    }
+    impl Default for ImportData {
+        fn default() -> Self {
+            Self {
+                re_exports: AstAlloc::vec(),
+                data: ImportTracker::default(),
+            }
+        }
+    }
+    pub(crate) type ImportToBind = ImportData;
+
+    pub struct ExportData {
+        pub potentially_ambiguous_export_star_refs: AstVec<ImportData>,
+        pub data: ImportTracker,
+    }
+    impl Default for ExportData {
+        fn default() -> Self {
+            Self {
+                potentially_ambiguous_export_star_refs: AstAlloc::vec(),
+                data: ImportTracker::default(),
+            }
+        }
+    }
+    pub(crate) type ResolvedExport = ExportData;
+
+    pub type RefImportData = ArrayHashMap<Ref, ImportData, AutoContext, AstAlloc>;
+    pub type ResolvedExports = StringArrayHashMap<ExportData, StringContext, AstAlloc>;
+    pub type ProbablyTypescriptType = ArrayHashMap<Ref, (), AutoContext, AstAlloc>;
+    pub type SortedAndFilteredExportAliases = AstVec<Box<[u8], AstAlloc>>;
+    pub type CjsExportCopies = AstVec<Ref>;
+    pub type TopLevelSymbolToParts = bun_ast::ast_result::TopLevelSymbolToParts;
+
+    #[derive(Clone, Copy, Default)]
+    pub struct Flags {
+        pub is_async_or_has_async_dependency: bool,
+        pub needs_exports_variable: bool,
+        pub force_include_exports_for_entry_point: bool,
+        pub needs_export_symbol_from_runtime: bool,
+        pub did_wrap_dependencies: bool,
+        pub needs_synthetic_default_export: bool,
+        pub wrap: WrapKind,
+    }
+    pub use crate::WrapKind as Wrap;
+
+    pub struct JSMeta {
+        pub probably_typescript_type: ProbablyTypescriptType,
+        pub imports_to_bind: RefImportData,
+        pub resolved_exports: ResolvedExports,
+        pub resolved_export_star: ExportData,
+        pub sorted_and_filtered_export_aliases: SortedAndFilteredExportAliases,
+        pub top_level_symbol_to_parts_overlay: TopLevelSymbolToParts,
+        pub cjs_export_copies: CjsExportCopies,
+        pub wrapper_part_index: Index,
+        pub entry_point_part_index: Index,
+        pub flags: Flags,
+    }
+
+    impl Default for JSMeta {
+        fn default() -> Self {
+            Self {
+                probably_typescript_type: ProbablyTypescriptType::default(),
+                imports_to_bind: RefImportData::default(),
+                resolved_exports: ResolvedExports::default(),
+                resolved_export_star: ExportData::default(),
+                sorted_and_filtered_export_aliases: AstAlloc::vec(),
+                top_level_symbol_to_parts_overlay: TopLevelSymbolToParts::default(),
+                cjs_export_copies: AstAlloc::vec(),
+                wrapper_part_index: Index::default(),
+                entry_point_part_index: Index::default(),
+                flags: Flags::default(),
+            }
+        }
+    }
+
+    bun_collections::multi_array_columns! {
+        pub trait JSMetaColumns for JSMeta {
+            probably_typescript_type: ProbablyTypescriptType,
+            imports_to_bind: RefImportData,
+            resolved_exports: ResolvedExports,
+            resolved_export_star: ExportData,
+            sorted_and_filtered_export_aliases: SortedAndFilteredExportAliases,
+            top_level_symbol_to_parts_overlay: TopLevelSymbolToParts,
+            cjs_export_copies: CjsExportCopies,
+            wrapper_part_index: Index,
+            entry_point_part_index: Index,
+            flags: Flags,
+        }
+    }
+
+    impl JSMeta {
+        pub type Flags = Flags;
+        pub type Wrap = crate::WrapKind;
+    }
+}
+
+pub use entry_point::EntryPoint;
+pub use js_meta::{
+    ExportData, ImportData, JSMeta, RefImportData, ResolvedExports, TopLevelSymbolToParts,
+};
 
 pub struct LinkerGraph<'a> {
     pub files: FileList,
